Point AI service calls to Railway, CORS fix, AdSet+Ad hierarchy

+++ src/main/java/com/adpilot/backend/campaign/Campaign.java
@@ -43,7 +43,7 @@ public class Campaign {
     private Double ctr;
     private Double cpc;
     private Double cpm;
-    private Integer conversions;
+    private Long conversions;
     private Double costPerResult;
     private Double roas;
